Work without clusters: require('euphoria').cluster(false); fixed sudden cessation of work worker`s

+++ index.js
@@ -1,42 +1,20 @@
 var cluster = require('cluster');
-var _extend  = require('util')._extend;
-var fs      = require('fs.extra');
+var fs      = require('fs-extra');
+var lib     = require('./lib/worker.js');
+var extend  = require('./lib/extend.js');
 
 var gracefull    = 0;
 var allready     = [];
 var dev          = (process.env.NODE_ENV && process.env.NODE_ENV == "dev");
+var useCluster   = true;
 
-function extend(obj, obj2) {
-	var keys = Object.keys(obj2);
-	keys.forEach(function(key){
-		if (obj[key]) {
-			if(typeof obj[key] == 'object' && typeof obj2[key] == 'object') {
-				var keys2 = Object.keys(obj2[key]);
-				keys2.forEach(function(key2){
-					if (obj[key][key2]) {
-						if(typeof obj[key][key2] == 'object' && typeof obj2[key][key2] == 'object') {
-							_extend(obj[key], obj2[key]);
-						} else {
-							obj[key][key2] = obj2[key][key2];
-						}
-					} else {
-						obj[key][key2] = obj2[key][key2];
-					}
-				});
-			} else {
-				obj[key] = obj2[key];
-			}
-		} else {
-			obj[key] = obj2[key];
-		}
-	});
-}
+
 
 
 function initWorker() {
 	var old = cluster.settings.exec;
 	cluster.setupMaster();
-	cluster.settings.exec = __dirname + "/worker.js";
+	cluster.settings.exec = __dirname + "/lib/worker.js";
 	var worker = cluster.fork();
 	cluster.settings.exec = old;
 	return worker;
@@ -69,32 +47,58 @@ function clusterExec(worker, func, args, callback) {
 
 
 function load(worker, name, object, callback){
-	clusterExec(worker, "load", [name, object], function(ret) {
-		extend(object, ret.result);
-		if (typeof callback == "function") callback();
+	process.nextTick(function () {
+		if (!useCluster) {
+			lib.load(name, object, callback);
+		} else {
+			clusterExec(worker, "load", [name, object], function(ret) {
+				if (ret.result) extend(object, ret.result);
+				if (typeof callback == "function") callback(ret.err, object);
+			});
+		}
 	});
 }
 
 
 function save(worker, name, object, callback){
-	clusterExec(worker, "save", [name, object], callback);
+	process.nextTick(function () {
+		if (!useCluster) {
+			lib.save(name, object, callback);
+		} else {
+			clusterExec(worker, "save", [name, object], function(ret) {
+				if (typeof callback == "function") callback(ret.err, object);
+			});
+		}
+	});
 }
 
 
 function _load(name, object, callback){
-	var worker = initWorker();
-	load(worker, name, object, function(){
-		worker.kill();
-		if (typeof callback == "function") callback(object);
+	process.nextTick(function () {
+		if (!useCluster) {
+			lib.load(name, object, callback);
+		} else {
+			var worker = initWorker();
+			load(worker, name, object, function(err, obj){
+				worker.kill();
+				if (typeof callback == "function") callback(err, obj);
+			});
+		}
 	});
 }
 
 
 function _save(name, object, callback){
-	var worker = initWorker();
-	save(worker, name, object, function() {
-		worker.kill();
-		if (typeof callback == "function") callback(object);
+	process.nextTick(function () {
+		if (!useCluster) {
+			lib.save(name, object, callback);
+		} else {
+			var worker = initWorker();
+			save(worker, name, object, function() {
+				worker.kill();
+				if (typeof callback == "function") callback(object);
+			});
+		}
 	});
 }
 
@@ -105,45 +109,52 @@ function setInterval(val){
 
 
 function connect(object, name, callback) {
+	process.nextTick(function () {
+		if (allready.indexOf(name) != -1)
+			throw new Error("[" + Date() + "] Database with the same name (" + name + ") is already connected!");
+		allready.push(name);
 
-	if (allready.indexOf(name) != -1)
-		throw new Error("Database with the same name (" + name + ") is already connected!");
-	allready.push(name);
+		fs.mkdirs(__dirname + "/db/");
 
-	fs.mkdirs(__dirname + "/db/");
+		var interval = 20000;
+		var worker = true;
+		if (useCluster)  worker = initWorker();
 
-	var interval = 5000;
-	var worker = initWorker();
+		var label = "[" + Date() + "] Euphoria: Database \"" + name + "\" is connected";
+		console.time(label);
+		load(worker, name, object, function(err) {
+			if (!err) {
+				function tick() {
+					if (worker) {
+						var saveStart = Date.now();
+						save(worker, name, object, function() {
+							interval = Math.max(20000, (Date.now()-saveStart)*10);
+							setTimeout(tick, interval);
+						});
+					} else {
+						console.warn("[" + Date() + "] Euphoria: Worker \"" + name + "\" not found!");
+					}
+				}
+				setTimeout(tick, interval);
 
-	var label = "DB.Euphoria: Database \"" + name + "\" is connected";
-	console.time(label);
-	load(worker, name, object, function() {
-
-		function tick() {
-			if (worker) {
-				var saveStart = Date.now();
-				save(worker, name, object, function() {
-					interval = Math.max(5000, (Date.now()-saveStart));
-					setTimeout(tick, interval);
-				});
+				gracefull++;
+				function gracefulExit() {
+					save(worker, name, object, function() {
+						console.log("Exit save for " + name + "_euphoria.db");
+						gracefull--;
+						if (useCluster) worker.kill();
+						worker = false;
+						if (!gracefull) process.exit();
+					});
+				}
+				process.once('SIGINT', gracefulExit).once('SIGTERM', gracefulExit);
+				console.timeEnd(label);
+			} else {
+				console.warn("[" + Date() + "] Euphoria: Database \"" + name + "\" is not connected (object offline)!");
 			}
-		}
-		setTimeout(tick, interval);
-
-		gracefull++;
-		function gracefulExit() {
-			save(worker, name, object, function() {
-				console.log("Exit save for " + name + "_euphoria.db");
-				gracefull--;
-				worker.kill();
-				if (!gracefull) process.exit();
-			});
-		}
-		process.once('SIGINT', gracefulExit).once('SIGTERM', gracefulExit);
-		console.timeEnd(label);
-		callback();
+			callback();
+		});
 	});
-
 }
 
 
@@ -155,3 +166,7 @@ module.exports.connect   = connect;
 module.exports.load      = _load;
 module.exports.save      = _save;
 module.exports.gracefull = gracefull;
+module.exports.cluster = function(bool){
+	useCluster = bool;
+	return module.exports;
+};
